Move test helpers to utils/ directory
